bodymeasure/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.5.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
Django==5.0.6
djangorestframework==3.15.1
dm-tree==0.1.8
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.1.1
etils==1.7.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.1
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
gast==0.5.4
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==4.33.0
gradio_client==0.17.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
immutabledict==4.2.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==20.9
pandas==2.1.4
pillow==10.3.0
promise==2.3
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.9
semantic-version==2.10.0
shellingham==1.5.4
simple_parsing==0.1.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-datasets==4.9.6
tensorflow-estimator==2.15.0
tensorflow-gcs-config==2.15.0
tensorflow-hub==0.16.1
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.15.0
tensorflow-probability==0.23.0
tensorflowjs==3.18.0
termcolor==2.4.0
tf-bodypix==0.4.2
tf_keras==2.16.0
tfjs-graph-converter==1.6.3
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
websockets==11.0.3
Werkzeug==3.0.3
wrapt==1.14.1
zipp==3.19.2

gradio==4.33.0
matplotlib==3.6.3
opencv_python==4.9.0.80
pandas==2.1.4

tf_bodypix==0.4.2
torch==2.3.0
tfjs_graph_converter